+++ CHANGELOG.md
@@ -5,6 +5,17 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## Next
+
+### Added
+
+- `searchcount` component
+- `selectioncount` component
+
+### Changed
+
+- `recording` primary highlight group to `Special` to make it a bit more eye catching by default
+
 ## [2.3.0] - 2025-07-11
 
 ### Added